Readjust modules to be more proper
