--- babel.config.js
@@ -1,13 +1,0 @@
-module.exports = {
-    presets: [
-        "@babel/preset-env",
-        "@babel/preset-react",
-        [
-            "@babel/preset-typescript",
-            {
-                isTSX: true,
-                allExtensions: true
-            }
-        ]
-    ]
-}
