+++ package/src/modules/component/index.ts
@@ -1,0 +1,1 @@
+export * from './glsl-pipeline-react'
